update scale infer
Notice that even use the same recipe, the scaling might be different, suspecting that it may be auto scaling for binary output. Try multiple method to find the header for the scale info but fail. Right now i'm using the high byte in the int part for scaling, can see that some of them still not quite right...